requirements.txt
requests
pySerial
mypy
flake8
black
types-requests
